Internal fsid is now 32 bits instead of 64

+++ unfs3/Config/exports.h
@@ -22,9 +22,9 @@ extern uint32   export_password_hash;
 extern unsigned char password[PASSWORD_MAXLEN+1];
 
 void		exports_parse(void);
-int		exports_options(const char *path, struct svc_req *rqstp, char **password, uint64 *fsid);
+int		exports_options(const char *path, struct svc_req *rqstp, char **password, uint32 *fsid);
 int             export_point(const char *path);
-char            *export_point_from_fsid(uint64 fsid);
+char            *export_point_from_fsid(uint32 fsid);
 nfsstat3	exports_compat(const char *path, struct svc_req *rqstp);
 nfsstat3	exports_rw(void);
